+++ packages/ui/src/components/molecules/index.ts
@@ -5,6 +5,12 @@ export {
 export { BackButton, type TBackButton } from './BackButton'
 export { CardProfile, type TCardProfile } from './CardProfile'
 export { CardHome, type TCardHome } from './CardHome'
+export {
+  CheckboxFilterWidget,
+  type TCheckboxFilterItem,
+  type TCheckboxFilterWidget,
+} from './CheckboxFilterWidget'
+export { CreateAuthor, type TCreateAuthor } from './CreateAuthor'
 export { InputGroup, type TInputGroup } from './InputGroup'
 export { Modal, type TModal } from './Modal'
 export { SelectGroup, type TSelectGroup } from './SelectGroup'