+++ packages/ui/CHANGELOG.md
@@ -2,11 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.18.0] - 2024-02-06
+## [0.20.0] - 2024-02-07
 
 ### Added
 
 - Migrate CardHome to UI
+
+## [0.19.0] - 2024-02-07
+
+### Added
+
+- Migrate CheckboxFilterWidget to UI
+
+## [0.18.0] - 2024-02-06
+
+### Added
+
+- Migrate CreateAuthor to UI
 
 ## [0.17.0] - 2024-02-06